+++ packages/custom_lint_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased patch
+
+- Improved formatting when specifying `source` on `encodePrioritizedSourceChanges`/`matcherNormalizedPrioritizedSourceChangeSnapshot`
+
 ## 0.5.12 - 2024-02-02
 
 - Added `encodePrioritizedSourceChanges`, to enable writing a `List<PrioritizedSourceChanges>` to a file